fix: исправил тип поля length в UARTDevice, добавил метод IsReady() для проверки готовности приема/передачи и пофиксил метод ReceiveLoRa

+++ src/Device/UARTDevice.cpp
@@ -7,11 +7,11 @@
 
 namespace IntroSatLib {
 
-    ISL_StatusTypeDef UARTDevice::WriteUART(uint8_t* buf, uint8_t nBytes, uint16_t timeout) {
+    ISL_StatusTypeDef UARTDevice::WriteUART(uint8_t* buf, uint16_t nBytes, uint16_t timeout) {
         return _uart.transmit(buf, nBytes, timeout);
     }
 
-    ISL_StatusTypeDef UARTDevice::ReadUART(uint8_t* buf, uint8_t nBytes, uint16_t timeout) {
+    ISL_StatusTypeDef UARTDevice::ReadUART(uint8_t* buf, uint16_t nBytes, uint16_t timeout) {
         return _uart.receive(buf, nBytes, timeout);
     }
